Allow mean/sd random interactions and mean work interactions to be double

+++ src/params.h
@@ -26,9 +26,9 @@ typedef struct{
 	int end_time;				    // maximum end time
 	int n_seed_infection;			// number of people seeded with the infections
 
-	int mean_random_interactions[N_AGE_TYPES]; // mean number of random interactions each day
-	int sd_random_interactions[N_AGE_TYPES];   // sd number of random interactions each day
-	int mean_work_interactions[N_WORK_NETWORKS];// mean number of regular work interactions
+	double mean_random_interactions[N_AGE_TYPES]; // mean number of random interactions each day
+	double sd_random_interactions[N_AGE_TYPES];   // sd number of random interactions each day
+	double mean_work_interactions[N_WORK_NETWORKS];// mean number of regular work interactions
 	double daily_fraction_work;      			// fraction of daily work interactions without social-distancing
 	double daily_fraction_work_used;      		// fraction of daily work interactions with social-distancing
 	double child_network_adults;				// fraction of adults in the child network
